03_AI_news_with_crewai/requirements.txt
python-dotenv
packaging
duckduckgo-search
fastapi
uvicorn
groq
request
pypdf
crewai
crewai[tools]
crewai_tools
pysbd
groq
langchain_openai